fix(web): star shader precision mismatch killed the whole graph view
u_correctionRatio was highp in vertex (GLSL default) but mediump in the
fragment shader -> WebGL program linking failed on first graph render ->
uncaught exception -> React unmounted #root (blank screen, prod incident
20.09). Fragment shader now declares precision highp float, matching
sigma's own NodeCircleProgram convention.

+++ web/src/lib/rendering.ts
@@ -127,7 +127,9 @@ void main() {
 `;
 
 const STAR_NODE_FRAGMENT_SHADER = /*glsl*/ `
-precision mediump float;
+// highp to match the vertex stage default: u_correctionRatio is shared by
+// both shaders, and a precision mismatch kills program linking (WebGL2).
+precision highp float;
 
 varying vec4 v_color;
 varying vec2 v_diffVector;
